add command line option '--checkLTNF/-k' to report atoms not in LTNF
The command line option '--checkLTNF/-k' is introduced in this commit.
If toggled, it prints atoms that are not in left-tuple normal
form to the top-level, with line number. An example:

Finding: Not in left-tuple normal form (LTNF): _John#_Student(+[_Mon _Tue _Fri] _coursehours+>20 _dept+>_Math _gender->_male -[1995 8 17]) at line 46