add support for C-style comment blocks, print deprecation warning for XML-style comment blocks
XML-style comment blocks (delimited by "<!--" and "-->") are now a
deprecated feature of the PSOA Presentation Syntax (PS) of
PSOATransRun 1.4.5 and are planned for removal in a future
release. Instead, the PS now uses the widely adopted C-style comment
blocks (delimited by "/*" and "*/"). The main reason for the change is
to strengthen the syntactic distinctiveness of both PSOA RuleML/PS and
PSOA RuleML/XML, also avoiding confusion of the syntaxes by software
tools.